backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
langchain==0.0.350
langchain-openai==0.0.2
python-dotenv==1.0.0
prometheus-client==0.19.0
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6

